requirements.txt
numpy>=1.23
pandas>=1.5
scikit-learn>=1.2
torch>=2.0
hyperopt>=0.2.7
shap>=0.42
matplotlib>=3.6
pyyaml>=6.0
scipy>=1.10
